requirements.txt
httpx>=0.27.0
fast-flights>=2.2
python-dotenv>=1.0.0
ruff>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
